remoe output_to_Fits from mask to abstract mask
