+++ autoarray/mask/abstract_mask.py
@@ -6,9 +6,10 @@ import numpy as np
 from pathlib import Path
 from typing import Dict, Union
 
+from autoconf.fitsable import output_to_fits
+
 from autoarray.abstract_ndarray import AbstractNDArray
 
-from autoarray import exc
 from autoarray import type as ty
 
 logging.basicConfig()
@@ -92,10 +93,38 @@ class Mask(AbstractNDArray, ABC):
     def dimensions(self) -> int:
         return len(self.shape)
 
-    def output_to_fits(self, file_path: Union[Path, str], overwrite: bool = False):
+    def output_to_fits(self, file_path, overwrite=False):
         """
-        Overwrite with method to output the mask to a `.fits` file.
+        Write the Mask to a .fits file.
+
+        Before outputting a 2D NumPy array mask, the array may be flipped upside-down using np.flipud depending on
+        the project config files. This is for Astronomy projects so that structures appear the same orientation
+        as `.fits` files loaded in DS9.
+
+        Parameters
+        ----------
+        file_path
+            The full path of the file that is output, including the file name and `.fits` extension.
+        overwrite
+            If `True` and a file already exists with the input file_path the .fits file is overwritten. If `False`, an
+            error is raised.
+
+        Returns
+        -------
+        None
+
+        Examples
+        --------
+        mask = Mask2D(mask=np.full(shape=(5,5), fill_value=False))
+        mask.output_to_fits(file_path='/path/to/file/filename.fits', overwrite=True)
         """
+        output_to_fits(
+            values=self.astype("float"),
+            file_path=file_path,
+            overwrite=overwrite,
+            header_dict=self.pixel_scale_header,
+            ext_name="mask"
+        )
 
     @property
     def pixels_in_mask(self) -> int:
