+++ autoarray/mask/mask_1d.py
@@ -1,12 +1,9 @@
 from __future__ import annotations
 
-from astropy.io import fits
 import logging
 import numpy as np
 from pathlib import Path
 from typing import Dict, List, Tuple, Union
-
-from autoconf.fitsable import output_to_fits
 
 from autoarray.mask.abstract_mask import Mask
 
@@ -178,49 +175,3 @@ class Mask1D(Mask):
         return {
             "PIXSCA": self.pixel_scales[0],
         }
-
-    @property
-    def hdu_for_output(self) -> fits.PrimaryHDU:
-        """
-        The mask as a HDU object, which can be output to a .fits file.
-
-        The header of the HDU is used to store the `pixel_scale` of the array, which is used by the `Array1D.from_hdu`.
-
-        This method is used in other projects (E.g. PyAutoGalaxy, PyAutoLens) to conveniently output the array to .fits
-        files.
-
-        Returns
-        -------
-        The HDU containing the data and its header which can then be written to .fits.
-        """
-        return array_1d_util.hdu_for_output_from(
-            array_1d=self.astype("float"), header_dict=self.pixel_scale_header
-        )
-
-    def output_to_fits(self, file_path: Union[Path, str], overwrite: bool = False):
-        """
-        Write the 1D mask to a .fits file.
-
-        Parameters
-        ----------
-        file_path
-            The full path of the file that is output, including the file name and .fits extension.
-        overwrite
-            If `True` and a file already exists with the input file_path the .fits file is overwritten. If `False`,
-            an error is raised.
-
-        Returns
-        -------
-        None
-
-        Examples
-        --------
-        mask = Mask1D(mask=np.full(shape=(5,), fill_value=False))
-        mask.output_to_fits(file_path='/path/to/file/filename.fits', overwrite=True)
-        """
-        output_to_fits(
-            values=self.astype("float"),
-            file_path=file_path,
-            overwrite=overwrite,
-            header_dict=self.pixel_scale_header,
-        )
